+++ plone/app/widgets/tests/test_dx.py
@@ -938,7 +938,8 @@ class RelatedItemsWidgetTests(unittest.TestCase):
 
     def test_widget(self):
         from plone.app.widgets.dx import RelatedItemsWidget
-        context = Mock(absolute_url=lambda: 'fake_url')
+        context = Mock(absolute_url=lambda: 'fake_url',
+                       getPhysicalPath=lambda: ['', 'site'])
         context.portal_properties.site_properties\
             .getProperty.return_value = ['SomeType']
         widget = RelatedItemsWidget(self.request)
@@ -957,8 +958,9 @@ class RelatedItemsWidgetTests(unittest.TestCase):
                     'searchAllText': u'Entire site',
                     'searchText': u'Search',
                     'separator': ';',
-                    'vocabularyUrl': '/@@getVocabulary?name='
+                    'vocabularyUrl': 'fake_url/@@getVocabulary?name='
                                      'plone.app.vocabularies.Catalog',
+                    'rootPath': '/site',
                 },
             },
             widget._base_args()
@@ -968,7 +970,8 @@ class RelatedItemsWidgetTests(unittest.TestCase):
         """The pattern_options value for maximumSelectionSize should
         be 1 when the field only allows a single selection."""
         from plone.app.widgets.dx import RelatedItemsFieldWidget
-        context = Mock(absolute_url=lambda: 'fake_url')
+        context = Mock(absolute_url=lambda: 'fake_url',
+                       getPhysicalPath=lambda: ['', 'site'])
         context.portal_properties.site_properties\
             .getProperty.return_value = ['SomeType']
         field = Choice(
@@ -989,7 +992,8 @@ class RelatedItemsWidgetTests(unittest.TestCase):
         from zope.schema.interfaces import ISource
         from zope.schema.vocabulary import VocabularyRegistry
 
-        context = Mock(absolute_url=lambda: 'fake_url')
+        context = Mock(absolute_url=lambda: 'fake_url',
+                       getPhysicalPath=lambda: ['', 'site'])
         context.portal_properties.site_properties\
             .getProperty.return_value = ['SomeType']
         field = List(
@@ -1008,7 +1012,7 @@ class RelatedItemsWidgetTests(unittest.TestCase):
         self.assertFalse('maximumSelectionSize' in patterns_options)
         self.assertEqual(
             patterns_options['vocabularyUrl'],
-            '/@@getVocabulary?name=foobar&field=selectfield',
+            'fake_url/@@getVocabulary?name=foobar&field=selectfield',
             )
 
     def test_converter_RelationChoice(self):